% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{close_duckdb}
\alias{close_duckdb}
\title{Close DuckDB Connection}
\usage{
close_duckdb(conn)
}
\arguments{
\item{conn}{A DuckDB connection object.}
}
\value{
Invisible NULL.
}
\description{
Closes a DuckDB connection.
}
\examples{
\dontrun{
conn <- use_duckdb()
# Use the connection...
close_duckdb(conn)
}

}
