% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_data_types.R
\name{optimize_data_types}
\alias{optimize_data_types}
\title{Optimize data types in a DuckDB table}
\usage{
optimize_data_types(conn, data_table_ref, log_table_ref)
}
\arguments{
\item{conn}{A DBI connection object}

\item{data_table_ref}{The name of the data table in database}

\item{log_table_ref}{The name of the log table in database}
}
\value{
NULL invisibly
}
\description{
Optimize data types in a DuckDB table
}
\details{
Optimizes column data types by analyzing content and converting to appropriate types
}
\keyword{internal}
