% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_duckdb}
\alias{remove_duckdb}
\title{Remove DuckDB Database}
\usage{
remove_duckdb(conn, dbname = "redcap.duckdb")
}
\arguments{
\item{conn}{A DuckDB connection object.}

\item{dbname}{Character string specifying the database file name.
Default is "redcap.duckdb".}
}
\value{
Invisible NULL.
}
\description{
Closes a DuckDB connection and removes the database file.
}
\examples{
\dontrun{
conn <- use_duckdb()
# Use the connection...
remove_duckdb(conn)
}

}
