% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_duckdb}
\alias{use_duckdb}
\title{Create DuckDB Connection}
\usage{
use_duckdb(dbname = "redcap.duckdb")
}
\arguments{
\item{dbname}{Character string specifying the database file name.
Default is "redcap.duckdb".}
}
\value{
A DuckDB connection object.
}
\description{
Creates a DuckDB connection to the REDCap database file.
}
\examples{
\dontrun{
conn <- use_duckdb()
# Use the connection...
remove_duckdb(conn)
}

}
